Refactor and clean up.
