+++ jdkmanager/ListCommand.java
@@ -26,6 +26,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
+import jdkmanager.client.JdkClient;
+import jdkmanager.client.Version;
+import jdkmanager.util.Environment;
 import picocli.AutoComplete;
 import picocli.CommandLine;
 
@@ -34,7 +37,7 @@ import picocli.CommandLine;
  */
 @CommandLine.Command(name = "list", description = "Lists information about available JDK's.",
         subcommands = AutoComplete.GenerateCompletion.class)
-class JdkList extends BaseCommand {
+class ListCommand extends BaseCommand {
 
     @CommandLine.Option(names = {"-l", "--local"}, description = "Lists the JDK's installed locally.")
     private boolean local;
@@ -43,53 +46,34 @@ class JdkList extends BaseCommand {
     Integer call(final JdkClient client) throws Exception {
         if (local) {
             final AtomicInteger exitCode = new AtomicInteger(0);
-            final Path workDir;
-            final int depth;
-            if (distributionSet) {
-                workDir = distributionDir();
-                depth = 1;
-            } else {
-                workDir = Environment.WORK_DIR;
-                depth = 2;
-            }
+            final Path workDir = Environment.WORK_DIR;
             if (Files.notExists(workDir)) {
                 print("No JDK's installed for %s", distribution);
                 return 0;
             }
-            try (Stream<Path> paths = Files.walk(workDir, depth)) {
+            try (Stream<Path> paths = Files.walk(workDir, 2)) {
                 paths.sorted()
                         .forEachOrdered(dir -> {
-                    if (Files.exists(dir.resolve("bin").resolve(Environment.isWindows() ? "java.exe" : "java"))) {
-                        try {
-                            final var status = client.getJavaInfo(dir);
-                            if (status.exitStatus() > 0) {
-                                printError("Failed to determine info for %s. Ensure this is a valid JDK directory.", dir);
-                                exitCode.set(status.exitStatus());
-                                return;
+                            if (Files.exists(dir.resolve("bin")
+                                    .resolve(Environment.isWindows() ? "java.exe" : "java"))) {
+                                try {
+                                    final var properties = javaInfo(dir);
+                                    print("%s %s (%s): %s", properties.getProperty("java.vendor"),
+                                            properties.getProperty("java.version"), properties.getProperty("java.version.date"), dir);
+                                } catch (IOException | InterruptedException e) {
+                                    printError("Failed to determine info for %s: %s", dir, e.getMessage());
+                                    exitCode.set(1);
+                                }
                             }
-                            final var properties = status.body();
-                            print("%s %s (%s): %s", properties.getProperty("java.vendor"),
-                                    properties.getProperty("java.version"), properties.getProperty("java.version.date"), dir);
-                        } catch (IOException | InterruptedException e) {
-                            printError("Failed to determine info for %s: %s", dir, e.getMessage());
-                            exitCode.set(1);
-                        }
-                    }
-                });
+                        });
                 return exitCode.get();
             }
         }
-        final var status = client.getVersions();
-        if (status.exitStatus() > 0) {
-            printError("Failed to determine the available JDK's: %s", status.rawData().get());
-            return status.exitStatus();
-        } else {
-            final var versions = status.body();
-            print("@|bold LTS Versions for %s|@", distribution);
-            versions.lts().forEach(listConsumer(true));
-            print("@|bold Versions for %s|@", distribution);
-            versions.available().forEach(listConsumer(false));
-        }
+        final var versions = client.getVersions();
+        print("@|bold LTS Versions for %s|@", distribution);
+        versions.lts().forEach(listConsumer(true));
+        print("@|bold Versions for %s|@", distribution);
+        versions.available().forEach(listConsumer(false));
         return 0;
     }
